requirements.txt
# Librería para medir el uso de memoria
memory-profiler==0.61.0 
# Librería para la interacción con BigQuery
google_cloud_bigquery==3.20.1